+++ .github/CHANGELOG.md
@@ -11,6 +11,7 @@
 10. [Game] Enhancement - Added events for high and low building cases
 11. [Game] Enhancement - added neutral OPEN_DAY event
 12. [Game] Bug - Fix bug where game would crash on leaderboard due to duplicate commas in leaderboard preferences file
+13. [Game] Enhancement - changed charity event to use event choices
 
 
 # Changes from v1.2.0 to v1.2.1